# Unmarshal using token decoder