@iancoleman 2021-01-17
Thanks for this, much appreciated. The code is much cleaner using `Token` and the performance is much improved. Thanks again.